Use a single root dir changedetection.py entry point Remove 'scripts' from pip setup.py Add changedetection.py to entry_points
